refactor: consolidate analyzers into analytics package
