+++ analytics/code_quality/base_analyzer.py
